src/raspberry_testing/telemetry_collector/requirements.txt
paho-mqtt>=2.1.0
psutil>=5.9.6
